requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
langdetect==1.0.9
gtts==2.4.0
SpeechRecognition==3.10.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
tavily-python==0.2.8